+++ tempest_fastapi_sdk/__init__.py
@@ -395,7 +395,7 @@ from tempest_fastapi_sdk.websockets import (
     make_websocket_router,
 )
 
-__version__: str = "0.113.0"
+__version__: str = "0.114.0"
 
 __all__: list[str] = [
     "BASE_COLUMN_ORDER",